// src/app/model/CartItems.model.ts
export class CartItems{
    id: number | undefined;
    productCode: string |undefined;
    name: string |undefined;
    userId: number | undefined;
    productId: number | undefined;
    price: number | undefined;
    quantity: number | undefined;
    totalPrice: number | undefined;
    image?: string;
}